Merged PR 15089: Adds redis cache support + b2c caching
This adds support for redis caching and initial b2c cache implementation

The cache service is a "Factory" which given a domain type will retrieve the correct cache store which can be used by the different services.

Only redis cache is implemented at the moment but this could actually be anything since the typing will be correct no matter which cache handler set.

Notes:
- npm install required for new modules
- docker redis available
- new environment variable is required for redis (azure connection string format)
- no cache invalidation, only 5 min TTL

Future work:
- implement b2c cache invalidation on user changes
  - lock, name changes, ...
- maybe add cache to userInformation to avoid queries all together