Merged PR 19086: Handle user roles in activate unit
When activating a unit there is list of users selected to be unlocked. Of this list:
- all users that were locked are unlocked
- all user roles that were locked are unlocked

It is important that we don't update all users of the list regardless of them being locked or not because we only update their updated_at and updated_by columns in the case where their status changes.