feat: create progress update accept more levels
By adding milestones to the progress updates there was a need to have
different progress updates apart from the simple one. This are the one
level and two levels that enables orgs to have a more personalized
progress update creation.

This commit adds the functionality to have three different types of
progress updates: simple, one level, and two level.