task: notify me handler and listener
With new events being sent to queue a handler that receives and knows
how to respond to different type of events was needed. This commit
introduces a base handler that can later be transformed into abstract
and extended by different types of events and all the things associated
with it.