Task/update innovations on ES index (#291)
* task: improve performance and add missing fields

* feat: add listener that updates es innovation doc

This queue-based function listens to messages on a queue and when they
happen it will update the information of a given innovation on the ES
index.

* feat: decorator to trigger innovation index update

Decorator that will be used on the endpoints that will trigger the
innovation update on the elastic search index. It was created to make it
easy and reusable to update the index (when a innovation update occurs).

* task: add decorator to endpoints that affect

Add ES update decorator to all endpoints that will affect if the
innovation need to be updated or not.

* task: add cron that reindexes every dawn

A cron that runs every dawn that recreates the index and ingests all the
documents. This is to make sure that the data is always up-to-date.

* task: add throw error on function to retry automatic

* refactor: minor change