Feat/elastic search (#296)
* task: elastic search mapping and ingestion (#288)

* task: elastic search mapping and ingestion

Introduces the elastic search integration. Introduces the first
iteration of mapping and ingestion of documents, alongside a structure
on how to use the elastic search.

* task: add es container and package

* refactor: create a shared ES env

* Task/get search innovations list (#290)

* task: improve performance and add missing fields

* task: add get search innovation list (WIP)

Introduces the first working version of elastic search innovation list.
Currently in a "raw" manner with a lot of missing TODOs, but provides a
working prototype for FE to integrate with.

* refactor: create a helper to build es query

* refactor: some minor code cleanup and changes

- Simplified some conditions.
- Removed duplicate name from the index.

* refactor: change from must to filters

- Filters should be filters instead of must to not affect the score.
- Minor fixes related with sort of location and name.

* refactor: add comments and cleanup

* refactor: move postHandlers position

* replace import query

* fix: support status filter logic changed

* refactor: general improvements

---------

Co-authored-by: Manuel Sousa <manuel.sousa@bjss.com>

* Task/update innovations on ES index (#291)

* task: improve performance and add missing fields

* feat: add listener that updates es innovation doc

This queue-based function listens to messages on a queue and when they
happen it will update the information of a given innovation on the ES
index.

* feat: decorator to trigger innovation index update

Decorator that will be used on the endpoints that will trigger the
innovation update on the elastic search index. It was created to make it
easy and reusable to update the index (when a innovation update occurs).

* task: add decorator to endpoints that affect

Add ES update decorator to all endpoints that will affect if the
innovation need to be updated or not.

* task: add cron that reindexes every dawn

A cron that runs every dawn that recreates the index and ingests all the
documents. This is to make sure that the data is always up-to-date.

* task: add throw error on function to retry automatic

* refactor: minor change

* fix: escape reserved chars from es query (#292)

* task: add search by email to admin users

* Task/add email and suggestion filter (#293)

* task: add search by email to admin users

* task: add suggestedOnly filter

* task: add throw of errors on ES integration service (#294)

* Task/translate (#295)

- adds translation from enum/keywords into elasticsearch so that the fields are searchable
- changes schema for document keywords to text+keyword
- makes search filters use keyword field

---------

Co-authored-by: Manuel Sousa <manuel.sousa@bjss.com>
Co-authored-by: Manuel Sousa - BJSS <115169817+manuelsousa-bjss@users.noreply.github.com>