feat: update index is done using redis
This commit optimized the reindex of documents from innovation
documents. The previous aproach using event based on queue was not
working since it could lead to multiple reindex of the same document in
the row. By using a redis set we make sure that we update once the
innovation every 30s.