Poc/notify me (#260)
* task: added base notify me entites and types

This commit provides the base for the other features to start being
build. This added two new entities:

- `NotifyMeSubscription` that saves the notify me subscriptions for a
  user role (e.g., if a QA wants to be notified everytime the support
  status changes to waiting)
- `NotificationSchedule` saves notification that are to be sent later
  (e.g., a user wants to be reminded in 6 months to check the innovation)

Besides the entity this commit already sets the base structure of the
types to be used including the event payloads and subscription
configurations.

* task: notify me handler and listener

With new events being sent to queue a handler that receives and knows
how to respond to different type of events was needed. This commit
introduces a base handler that can later be transformed into abstract
and extended by different types of events and all the things associated
with it.

* task: subscribe, unsubscribe and list of notify me subscriptions

The user needed a way to manage the subscriptions to notify me.
This commit adds functionality for the user to be able to create, delete
and list his subscriptions.

* task: add notification scheduler cron

With the notify me a structure to save scheduled notifications was
needed and added. In order to handle this notifications that are
scheduled a cron is needed to check periodically if there are any
pending notifications to be sent.

* comment (#261)

* comment

* feat: improve the Joi error handling on cron

---------

Co-authored-by: Diogo M. Costa <diogo.costa@bjss.com>

---------

Co-authored-by: Manuel Sousa - BJSS <115169817+manuelsousa-bjss@users.noreply.github.com>
Co-authored-by: Manuel Sousa <manuel.sousa@bjss.com>