Task/persist schema in memory (#346)
* feat: persist the schema in memory

Introduces a singleton service that handles the load of the schema in
order to reduce the DB traffic. Since the schema will be used across
multiple places this optimization was needed. Taking into consideration
this change other places were changed that were using the old methods.

* fix: errors found in the sync with FE

* task: change id validation

A decision was made to let all types of a string be ids since the client
will not change directly the ids.