task: hasFiles directly on schema (#363)
Sections that can upload files were hardcoded in the code, by having a
schema that defined the IR rules this needed to be changed in order to
be trully dynamic. This commit adds a property `hasFiles` to the
subsections that can be used to defined if it accepts file uploads while
is not shared or not.